Update open-your-source-as-if-you-gonna-die-tonight.md
In the sprit of open source - I've made some improvements: fixed some typos, shifted some of the sentences around a bit. Added a few links.

Changed "staff" to "stuff" a lot ;) (staff is a person that works somewhere, or a long pole. Stuff is a word that describes assorted things. Like, a "pile of stuff")